Prevent spurious test failure in testShuttingDownFailsRegistration (#338)
Motivation:

There were two errors in testShuttingDownFailsRegistration that
could cause spurious test failures.

The first was that the test did not wait for connection establishment
before it began shutting the selector down, due to a change made in

The second was that the server side of the connection was not wedged open,
and so it would be closed by the call to loop.closeGently. That closure
could, if the main thread took long enough, lead to the client side
reading EOF and being forcibly closed.

Modifications:

Added a promise to wait for connection establishment.
Added support for wedging both the server and client side open.

Result:

Fewer test failures.