add HTTPURI for more performant uri access Motivation:
Currently the only way to access a request head's uri is via a computed property that converts it to a String. The internal storage however is usually a byte buffer. Being able to access the uri bytes without going through String would be a performance win.

Modifications:

Removed URI in favor or HTTPURI and internal HTTPURIStorage

Result:

HTTP URI can be accessed more performantly.