Use `Int` to store nanosecs on 64b and `Int64` on 32b
Motivation:

`Int64` is the right choice here, but it breaks SemVer
and needs to wait for NIO 2.x. So instead we just use
`Int64` on ARM 32-bit, and stick w/ `Int` on the other
platforms.

Modifications:

Added a typealias `TimeAmount.Value` which aliases to
`Int` on 64-bit platforms and `Int64` on arch(arm).

Result:

Result should be SemVer minor.