Change initializer to public for HTTPRequestDecoder (#438)
Motivation:

d53ee6dafc4d03a342aa32e79d5d282abc0c4f1a introduced a new constructor to HTTPRequestDecoder which allows to change the behaviour of handling left over bytes when an upgrade was detected and the decoder was removed. This was done as an internal init block as we wanted to to do as part of a patch release.

Modifications:

Change internal to public init

Result:

More flexible configuration possible.