Updated the code comment to explain the correct type of Future result.
Motivation:

This is done because when using Xcode and option + Click on a function you get a help overview inline to see more information and the messaging was incorrect. This helps developers in the future who want to rely on docs via Xcode.

Modification:

Correct the docs

Result:

More correct docs.