silence warning
Motivation:

Swift has a rather unhelpful warning

    warning: treating a forced downcast to 'EmbeddedEventLoop' as optional will never produce 'nil'

when assigning that to a variable of type `EmbeddedEventLoop!`. The
warning is accurate but obviously we know that can never be `nil`. The
way to silence this warning is to put parenthesis around the expression.

Modifications:

added parenthesis around force casts that are assigned to IUOs

Result:

fewer warnings with 4.2