Nil NIOHTTP1Server dynamic handlers
Motivation:

The server attaches a dynamic handler to requests and will keep passing
data until the handlers are removed; which never happens.

Modifications:

`nil` out the `handler` when `completeResponse` is called.

Result:

When connections are reused we now get the correct dynamic handler for
it instead of the previous one.