Support user defined writability in Channels
Mimics the user defined writability feature present in Netty’s `ChannelOutboundBuffer`.  It adds the feature to `PendingWritesManager` for both stream & datagram sockets. Instead of exposing a reference to a `PendingWritesManager` in `Channel.unsafe` (as Netty does with `ChannelOutboundBuffer`) the `Channel` options feature (`getOption` & `setOption`) is used to allow setting the user defined writbability of the `Channel` (which is forwarded to the `PendingWritesManager` internally).