Iterating over the EventLoops from EventLoopGroups (#509)
* EventLoopIterator

Provide a way to iterate over the EventLoops that compose an EventLoopGroup

Motivation:
For some use cases, it might be necessary to access the underlying event loops. See #497.

Modifications:
Added an EventLoopIterator as well as a new protocol requirement to EventLoopGroup called makeIterator() -> EventLoopIterator? . This is optional to avoid a SemVer Major, but will be made non-optional with 2.0.0. A current implementation is provided for every EventLoop through a protocol extension as well as for the MultiThreadedEventLoopGroup.

Result:
It is now possible to iterate over the EventLoops that compose a EventLoopGroup