Remove a stray space from the end of an error message
Motivation:

The stray space looks like an oversight.

Modifications:

Removed a stray space from the end of a string interpolation literal
that's used to construct an error message.

Result:

The exact value (but not the meaning) of the error message string
changes. Clients that match against the exact error message might break,
but that's arguably bad practice.