run a netstat if there's a socket leak in integration tests
Motivation:

We have a flaky integration test (#563) and I can't reproduce locally in
docker unfortunately. We have an `lsof` output but it just reads

    NIOHTTP1S 5157 root   12u  sock       0,7      0t0 113688082 can't identify protocol

which isn't very helpful. That's a known `lsof` on Linux but and the
recommendation is to run `netstat` which we now do.

Modifications:

add netstat output to the integration tests debugging info.

Result:

hopefully more information so we can soon fix #563