unify dead pointer dummy values
Motivation:

Previously we used 0xdeadbeef for all dead pointers but that doesn't fit
into an `Int` on 32-bit platforms.

Modifications:

change the dead pointer value to 0xdeadbee which is good because even
after offsetting the pointer by a couple of bytes, we'll still have
'dead' in the registers which should give us a hint.

Result:

- unified dead pointer values on 32 and 64 bit