Work around SE-0213
Motivation:

In Swift 5, types that conform to ExpressibleByIntegerLiteral will now
have their init(integerLiteral:) initializer invoked whenever the
compiler sees code of the form `Type(integerLiteral)`, rather than
before where the integer literal would be created as `Int` and then
an alternative initializer would be called. This broke our tests, so
we should fix it now that we're aware of it.

Modifications:

Force the type of the argument for the one broken test line, add new
test lines that execute this code path.

Result:

Better testing, fewer compile errors.