connect timeout tests: work around ECONNREFUSED (#653)
Motivation:

Some network configurations might disallow connecting to unknown IP
addresses. The effect of that is that we might see `ECONNREFUSED` when
trying to connect to 198.51.100.254 which nominally is reserved for
documentation only (ie. we shouldn't get any RSTs).

Modifications:

ignore `ECONNREFUSED` in the timeout tests too

Result:

tests stable in less permissive network environments